Translation Alignments: Port translation alignment endpoints from Flask app to ATLAS

+++ readhomer_atlas/library/importers/versions.py
@@ -11,8 +11,8 @@ from treebeard.exceptions import PathOverflow
 
 from readhomer_atlas import constants
 
-from .models import Node
-from .urn import URN
+from ..models import Node
+from ..urn import URN
 
 
 LIBRARY_DATA_PATH = os.path.join(settings.PROJECT_ROOT, "data", "library")
@@ -268,8 +268,9 @@ def get_first_value_for_language(values, lang):
     return next(iter(filter(lambda x: x["lang"] == lang, values)), None).get("value")
 
 
-def import_versions():
-    Node.objects.filter(kind="nid").delete()
+def import_versions(reset=True):
+    if reset:
+        Node.objects.filter(kind="nid").delete()
 
     library = resolve_library()
